+++ changelog.md
@@ -6,6 +6,7 @@
 + Implement pages bookmarks. This feature allows users to bookmark pages for faster and better navigation.
 + Implement all-in-one.Dockerfile, a Docker or Podman cofiguration file, for streamlining self hosting MWiki. The all-in-one image provides a multi-process Linux container with MWiki and a pre-configured Caddy reverse proxy and static web server. This approach simplifies self-hosting as all Caddy configuration can be tuned via environment variables and the user does not need to take care of multiple containers. While this approach is not scalable, it is good enough for a deployment on a single machine.
 + Change footnote syntax in order to avoid parsing ambiguity. The MWiki syntax for footnote was changed from `^{footnote here}` to `^{@ footnote here @}`. The Obsidian's compatibility syntax `^[footnote here]` for footnotes still remains.
++ Implement command $ mwiki manage --shell for flask db shell.
 
 ## Release v0.9.3